keylogger_app/requirements.txt

pynput>=1.7.7
Pillow>=10.4.0
cryptography>=41.0.7
pygame>=2.6.0
psutil>=6.0.0
pyperclip>=1.9.0
pyautogui>=0.9.54
